requirements.txt
fire==0.5.0
matplotlib==3.5.0
numpy==1.23.5
pandas==1.2.4
PyYAML==6.0
scipy==1.8.0
torch==2.0.0
